refactor(android): AlarmFortuneSettings(1005) 재분할 → +AlarmSnoozeSettings.kt
음성번역/알람 row/스누즈 설정을 별도 파일로 분리. 두 파일 모두 1000줄 미만.
